+++ Backend/src/Meeting/Spotly.Meeting.Domain/Entities/MeetingTag.cs
@@ -1,4 +1,6 @@
-﻿namespace Spotly.Meeting.Domain.Aggregates;
+﻿using Spotly.Meeting.Domain.ValueObjects;
+
+namespace Spotly.Meeting.Domain.Entities;
 
 /// <summary>
 /// Сущность, описывающая тег встречи.
