+++ Backend/src/Meeting/Spotly.Meeting.Domain/ValueObjects/EndAtDate.cs
@@ -1,4 +1,4 @@
-namespace Spotly.Meeting.Domain.Aggregates;
+namespace Spotly.Meeting.Domain.ValueObjects;
 
 /// <summary>
 /// Объект-значение, которое представляет дату и время окончания встречи.
