+++ Backend/src/Meeting/Spotly.Meeting.Domain/ValueObjects/MeetingTitle.cs
@@ -1,7 +1,7 @@
 using System.Text.RegularExpressions;
 using Spotly.Meeting.Domain.Exceptions;
 
-namespace Spotly.Meeting.Domain.Aggregates;
+namespace Spotly.Meeting.Domain.ValueObjects;
 
 /// <summary>
 /// Объект-значение, который представляет собой заголовок встречи.
